+++ PrintingExample/src/main/java/com/aevi/example/print/PrintingActivity.java
@@ -2,7 +2,7 @@ package com.aevi.example.print;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.util.SparseArray;
 import android.view.View;
@@ -30,6 +30,8 @@ import butterknife.BindViews;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 import butterknife.OnItemSelected;
+import butterknife.Setter;
+import butterknife.ViewCollections;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.annotations.NonNull;
 import io.reactivex.disposables.Disposable;
@@ -74,12 +76,7 @@ public class PrintingActivity extends AppCompatActivity {
     @BindView(R.id.codepages)
     ViewGroup codepagesLayout;
 
-    static final ButterKnife.Setter<View, Boolean> ENABLED = new ButterKnife.Setter<View, Boolean>() {
-        @Override
-        public void set(View view, Boolean value, int index) {
-            view.setEnabled(value);
-        }
-    };
+    static final Setter<View, Boolean> ENABLED = (view, value, index) -> view.setEnabled(value);
 
     private static final int MAX_STATUS_LENGTH = 3;
 
@@ -297,7 +294,7 @@ public class PrintingActivity extends AppCompatActivity {
 
     private void enableButtons(boolean enable) {
         if (enable) {
-            ButterKnife.apply(buttons, ENABLED, enable);
+            ViewCollections.set(buttons, ENABLED, enable);
         }
     }